seperate auth controller
