+++ src/controller/userController.ts
@@ -42,50 +42,19 @@ export const getAllUsers = (
   }
 };
 
-export const loginUser = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<void> => {
-  try {
-    const { email, password } = req.body;
-    logger.info("User login attempt", { email });
-    const user = await userService.validateUserCredentials(email, password);
-    if (!user) {
-      return next(
-        new ApiError(StatusCodes.UNAUTHORIZED, "Invalid email or password")
-      );
-    }
-    const tokens = userService.generateTokens(user);
-    res.status(StatusCodes.OK).json(tokens);
-  } catch (error) {
-    logger.error("Failed to log in user", { error });
-    next(
-      new ApiError(StatusCodes.INTERNAL_SERVER_ERROR, "Failed to log in user")
-    );
-  }
-};
-
 export const createNewUser = async (
   req: Request,
   res: Response,
   next: NextFunction
-): Promise<void> => {
+) => {
   try {
     const { name, email, password } = req.body;
     logger.info("Creating User");
-    if (userService.fetchUserByEmail(email)) {
-      return next(
-        new ApiError(StatusCodes.BAD_REQUEST, "Email already in use")
-      );
-    }
     const newUser = await userService.createUser(name, email, password);
     res.status(StatusCodes.CREATED).json(newUser);
   } catch (error) {
     logger.error("Failed to create user");
-    next(
-      new ApiError(StatusCodes.INTERNAL_SERVER_ERROR, "Failed to create user")
-    );
+    next(error);
   }
 };
 
@@ -98,13 +67,7 @@ export const updateUser = async (
     const userId = parseInt(req.params.id);
     const updateData = req.body;
     logger.info("Updating user", { userId, updateData });
-
     const updatedUser = await userService.updateUser(userId, updateData);
-    if (!updatedUser) {
-      return next(
-        new ApiError(StatusCodes.NOT_FOUND, `User with ID ${userId} not found`)
-      );
-    }
     res.status(StatusCodes.OK).json(updatedUser);
   } catch (error) {
     logger.error("Failed to update user", { error });
@@ -121,43 +84,9 @@ export const deleteUser = async (
     const userId = parseInt(req.params.id);
     logger.info("Deleting user", { userId });
     const deletedUser = userService.deleteUser(userId);
-    if (!deletedUser) {
-      return next(
-        new ApiError(StatusCodes.NOT_FOUND, `User with ID ${userId} not found`)
-      );
-    }
     res.status(StatusCodes.OK).json(deletedUser);
   } catch (error) {
     logger.error("Failed to delete user", { error });
     next(error);
-  }
-};
-
-export const refreshToken = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): void => {
-  try {
-    const { refreshToken } = req.body;
-    logger.info("Refreshing token");
-    if (!refreshToken) {
-      return next(
-        new ApiError(StatusCodes.BAD_REQUEST, "Refresh token is required")
-      );
-    }
-    const tokens = userService.refreshAccessToken(refreshToken);
-    if (!tokens) {
-      return next(new ApiError(StatusCodes.FORBIDDEN, "Invalid refresh token"));
-    }
-    res.status(StatusCodes.OK).json(tokens);
-  } catch (error) {
-    logger.error("Failed to refresh access token", { error });
-    next(
-      new ApiError(
-        StatusCodes.INTERNAL_SERVER_ERROR,
-        "Failed to refresh access token"
-      )
-    );
   }
 };
